@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.4.0",
         "ext-json": "*"
     },
     "autoload": {
